+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -3,8 +3,12 @@ package ru.skypro.homework.service.impl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 import org.webjars.NotFoundException;
 import ru.skypro.homework.models.dto.AdsDto;
 import ru.skypro.homework.models.dto.CreateAdsDto;
@@ -44,7 +48,8 @@ public class AdsServiceImpl implements AdsService {
     @Override
     public AdsDto addAds(CreateAdsDto ads, Images images) {
         log.info("Trying to add new ad");
-        User user = userService.getUser(0);//todo заменить на текущего авторизованного юзера
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUser(authentication.getName());
         Ads newAds = adsMapper.fromCreateAds(ads, user, images);
         Ads response = adsRepository.save(newAds);
         log.info("The ad with pk = {} was saved ", response.getPk());
@@ -54,17 +59,24 @@ public class AdsServiceImpl implements AdsService {
 
     @Override
     public List<AdsDto> getAdsMe(Boolean authenticated, String authority, Object credentials, Object details, Object principal) {
-        // FIXME: Just returns all
         log.info("Trying to get all user's ads");
-
-        return toAdsDtoList(adsRepository.findAll());
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUser(authentication.getName());
+        List<Ads> list = adsRepository.findAll();
+        return toAdsDtoList(list.stream().filter(e -> e.getAuthor().equals(user)).collect(Collectors.toList()));
     }
 
     @Transactional
     @Override
     public void removeAds(Integer id) {
         log.info("Trying to remove the ad with id = {}", id);
-        getAds(id);
+        Ads ads = getAds(id);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUser(authentication.getName());
+        if(!ads.getAuthor().equals(user) && !userService.isAdmin(authentication)) {
+            log.warn("Unavailable to remove. It's not your ads! ads author = {}, username = {}", ads.getAuthor().getEmail(), user.getEmail());
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Unavailable to remove. It's not your ads!");
+        }
         adsRepository.deleteById(id);
         log.info("The ad with id = {} was removed", id);
     }
@@ -81,7 +93,13 @@ public class AdsServiceImpl implements AdsService {
     public AdsDto updateAds(Integer id, CreateAdsDto adsDto) {
         log.info("Trying to update the ad with id = {}", id);
         Ads ads = getAds(id);
-        User user = userService.getUser(0);//todo использовать текущего юзера
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = userService.getUser(authentication.getName());
+        if(!ads.getAuthor().equals(user) && !userService.isAdmin(authentication)) {
+            log.warn("Unavailable to update. It's not your ads! ads author = {}, username = {}", ads.getAuthor().getEmail(), user.getEmail());
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Unavailable to update. It's not your ads!");
+        }
+
         Ads updatedAds = adsMapper.fromCreateAds(adsDto, user, ads.getImage());
         updatedAds.setComments(List.copyOf(ads.getComments()));
         updatedAds.setPk(id);
@@ -125,6 +143,8 @@ public class AdsServiceImpl implements AdsService {
     }
 
     private List<AdsDto> toAdsDtoList(List<Ads> ads) {
-        return ads.stream().map(adsMapper::toAdsDto).collect(Collectors.toList());
+        return ads.stream()
+                .map(adsMapper::toAdsDto)
+                .collect(Collectors.toList());
     }
 }
